requirements.txt
# -----------------------------------------------------------------------
# AIHawk Jobs Applier – Python dependencies
# -----------------------------------------------------------------------

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0

# Task queue (Celery + in-memory broker for single-user mode)
celery>=5.4.0
# Optional: install redis extra for production broker
# celery[redis]>=5.4.0

# Resume builder (upstream library)
git+https://github.com/feder-cr/lib_resume_builder_AIHawk.git

# LLM / AI
langchain==0.2.11
langchain-anthropic
langchain-huggingface
langchain-community==0.2.10
langchain-core==0.2.36
langchain-google-genai==1.0.10
langchain-ollama==0.1.3
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
openai==1.37.1

# HTTP
httpx~=0.27.2

# Browser automation (PDF generation)
selenium==4.9.1
webdriver-manager==4.0.2
undetected-chromedriver==3.5.5

# Data / validation
pydantic>=2.0.0
PyYAML~=6.0.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1

# CLI
click
inquirer
inputimeout==1.0.4

# Logging
loguru==0.7.2

# Utilities
python-dotenv~=1.0.1
regex==2024.7.24
Levenshtein>=0.26.0
reportlab==4.2.2
pdfminer.six==20221105

# Testing
pytest>=8.3.3
pytest-mock
pytest-cov